requirements.txt
Django==1.10.1
configparser==3.5.0
crispy-forms-foundation==0.5.3
django-crispy-forms==1.5.2
enum34==1.1.6
flake8==3.0.4
jedi==0.9.0
jsonfield==1.0.3
mccabe==0.5.2
pep8==1.7.0
pycodestyle==2.0.0
pyflakes==1.2.3
wheel==0.24.0
